move and update classes to new architecture
